many fixes

+++ SprintLevelEditor/SimpleRectangle.cs
@@ -22,10 +22,10 @@ namespace SprintLevelEditor
             Height = height;
         }
 
-        public static SimpleRectangle fromWall(Wall wall, int blockSize, int xOffset, int yOffset)
+        public static SimpleRectangle fromWall(Wall wall, float blockSize, int xOffset, int yOffset)
         {
             Rectangle rectangle = new Rectangle((int)wall.sprite.position.X, (int)wall.sprite.position.Y, (int)wall.sprite.DrawSize.Width, (int)wall.sprite.DrawSize.Height);
-            return new SimpleRectangle(Math.Max(0, (rectangle.X - xOffset) / blockSize), Math.Max(0, (rectangle.Y - yOffset) / blockSize), Math.Max(1, rectangle.Width / blockSize), Math.Max(1, rectangle.Height / blockSize));
+            return new SimpleRectangle(Math.Max(0, (int) ((rectangle.X - xOffset) / blockSize)), Math.Max(0, (int) ((rectangle.Y - yOffset) / blockSize)), Math.Max(1, (int) (rectangle.Width / blockSize)), Math.Max(1, (int) (rectangle.Height /  blockSize)));
         }
     }
 }
